feat: print

+++ src/includes_h/ast.h
@@ -54,7 +54,7 @@ ASTNode* parse_expression(Token **tokens);
 ASTNode* parse_term(Token **tokens);
 ASTNode* parse_primary(Token **tokens);
 ASTNode* parse_assignment(Token **tokens);
-int eval(ASTNode *node);
+float eval(ASTNode *node);
 void free_ast(ASTNode *node);
 
 #endif //INTERPRETEURC_AST_H
